% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisREAD.R
\name{cisREAD}
\alias{cisREAD}
\title{Performs the cisREAD workflow in one step upon a list of genes, from ATAC-seq , RNA-seq and TF footprint inputs
This function is a wrapper of map_peaks2genes, find_coCREs and select_coCREs which can be iterated over a list
of genes.}
\usage{
cisREAD(
  peaks,
  TFprofile,
  RNA,
  genes,
  genome = "hg38",
  distance = 1e+05,
  minTFevents = 2,
  coCREcutoff = 0.3,
  weights = c(1, 1),
  lambda = c("lambda_min", "lambda_se"),
  pval = 0.05,
  padj = 0.05
)
}
\arguments{
\item{peaks}{An ATAC-seq count matrix, where the first 6 columns are in BED
format, and columns afterwards are samples. Each row of the matrix should
give normalised, log2 transformed ATAC-seq counts for each ATAC-seq peak.}

\item{TFprofile}{A binary matrix where each row is an ATAC-seq peak, and each column indicates the presence (1)
or absence (0) of a TF footprint in a given differentiation stage}

\item{RNA}{A gene expression matrix where each row is a gene's ENSEMBL ID (if not then please convert first)
and each column gives the gene's expression (log2 normalised counts) in a given sample}

\item{genes}{A character vector of protein-coding gene symbols to find and select coCREs for -
e.g. c("PRDM1", "MYC", "IRF4","BCL6")}

\item{genome}{The genome annotation, only "hg38" is currently supported}

\item{distance}{The maximum genomic distance allowed between a candidate CRE and the gene's TSS. Cannot be
larger than the distance used in maps_peaks2genes}

\item{minTFevents}{Used to filter out candidate CREs not bound by any TFs. All candidate CREs must have at least
this number of 1's in the TFprofile matrix (default 2)}

\item{coCREcutoff}{The minimum similarity score (TF footprint dice similarity * ATAC-seq pearson correlation)
used to draw an edge between two candidate CREs in network construction. Should be a value between 0 and 1
(default 0.3), ower values will result in looser communities and higher values in tighter communities}

\item{weights}{A vector giving the weights assigned to TF similarity (1st value) and chromatin accessibility
correlation (2nd value). The defualt - c(1,1) - gives equal weighting to both, using c(0,1) would perform
community detection solely on chromatin accessibility}

\item{lambda}{The value of lambda used for model selection during cross-validation, one of
"lambda_min" (default) and "lambda_se". "lambda_min" will construct the model with the mean
minimum cross-validated squared error, "lambda_se" will construct the model within 1 standard error
of "lambda_min", and will result in a sparser model}

\item{pval}{The p-value alpha level for assigning statistical significance. coCRE predictors with
p < this value will be considered statistically significant}

\item{padj}{The Benjamini-Hockberg adjusted p-value alpha level for assigning statistical significance to
LASSO models. Allows for multiple testing correction when applying cisREAD to many genes.
(co)CRE predictors with p < pval, from gene-specific models with p < padj will be considered statistically
significant}
}
\value{
A cisREADResult object, listing significant, selected and all candidate CRE communities
}
\description{
Performs the cisREAD workflow in one step upon a list of genes, from ATAC-seq , RNA-seq and TF footprint inputs
This function is a wrapper of map_peaks2genes, find_coCREs and select_coCREs which can be iterated over a list
of genes.
}
