% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coCREs.R
\name{find_coCREs}
\alias{find_coCREs}
\title{Find communities of Cis-Regulatory-Elements (coCREs)}
\usage{
find_coCREs(
  peaks2genes,
  TFprofile,
  gene,
  distance = 1e+05,
  minTFevents = 2,
  coCREcutoff = 0.3,
  coCREgroupings = c("integrated", "ATAC", "TF")
)
}
\arguments{
\item{peaks2genes}{A dataframe mapping ATAC-seq peaks to nearby genes. Can be created with cisREAD's
map_peaks2genes function}

\item{TFprofile}{A binary matrix where each row is an ATAC-seq peak, and each column indicates the presence (1)
or absence (0) of a TF footprint in a given differentiation stage}

\item{gene}{A string giving a protein coding gene symbol, e.g. "PRDM1" or "MYC"}

\item{distance}{The maximum genomic distance allowed between a candidate CRE and the gene's TSS. Cannot be
larger than the distance used in maps_peaks2genes}

\item{minTFevents}{Used to filter out candidate CREs not bound by any TFs. All candidate CREs must have at least
this number of 1's in the TFprofile matrix (default 2)}

\item{coCREcutoff}{The minimum similarity score (TF footprint dice similarity * ATAC-seq pearson correlation)
used to draw an edge between two candidate CREs in network construction. Should be a value between 0 and 1
(default 0.3), lower values will result in looser communities and higher values in tighter communities}

\item{coCREgroupings}{Whether to perform community detection based  on ATAC-seq correlation ("ATAC"),  TF
binding similarity ("TF"), or integrated similarity ("integrated") - defined as ATAC-seq correlation * TF similarity
(default = "integrated"). Using "ATAC" or "TF" will result in detection of looser communities than "integrated",
it is therefore recommended to raise the "coCREcutoff" parameter to compensate.}
}
\value{
A find_coCREsResult object, listing candidate CRE communities, their TFs, and accessibilites
}
\description{
This function finds communities of co-accessible and co-bound cis-regulatory elements using infomap
community detection.
}
