% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_peaks2genes.R
\name{map_peaks2genes}
\alias{map_peaks2genes}
\title{Map ATAC-seq peaks to nearby genes
This function maps accessible genomic regions to protein coding genes
within a given distance.}
\usage{
map_peaks2genes(peaks, genome = "hg38", distance = 1e+05)
}
\arguments{
\item{peaks}{An ATAC-seq count matrix, where the first 6 columns are in BED
format, and columns afterwards are samples. Each row of the matrix should
give normalised, log2 transformed ATAC-seq counts for each ATAC-seq peak.}

\item{genome}{The genome annotation, only "hg38" is currently supported}

\item{distance}{The maximum genomic distance, in base pairs, allowed between a
gene's TSS and the centre of the ATAC-seq peak (default = 100000)}
}
\value{
A dataframe where each peak is mapped to all genes within
the given distance
}
\description{
Map ATAC-seq peaks to nearby genes
This function maps accessible genomic regions to protein coding genes
within a given distance.
}
