% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_coCREs.R
\name{select_coCREs}
\alias{select_coCREs}
\title{Selects cis-regulatory elements and cis-regulatory communities to regulate a gene}
\usage{
select_coCREs(
  coCREs,
  RNA,
  lambda = c("lambda_min", "lambda_se"),
  pval = 0.05,
  plot = FALSE
)
}
\arguments{
\item{coCREs}{A find_coCREsResult object created with cisREAD's find_coCREs function}

\item{RNA}{A gene expression matrix where each row is a gene's ENSEMBL ID (if not then please convert first)
and each column gives the gene's expression (log2 normalised counts) in a given sample}

\item{lambda}{The value of lambda used for model selection during cross-validation, one of
"lambda_min" (default) and "lambda_se". "lambda_min" will construct the model with the mean
minimum cross-validated squared error, "lambda_se" will construct the model within 1 standard error
of "lambda_min", and will result in a sparser model}

\item{pval}{The p-value cutoff for assigning statistical significance. coCRE predictors with
p < this value will be considered statistically significant}

\item{plot}{Indicates whether to construct glmnet plots during LASSO cross-validation, will visualise the
relationship between lambda and the number of co-efficients selected (default = FALSE)}
}
\value{
A select_coCREsResult object, listing selected CRE communities, their TFs and accessibilities, alongside
the glmnet model object, inputs and coefficients
}
\description{
This function finds CREs and coCREs whose chromatin accessibility best predicts the expresion of a target
gene using LASSO regression. This should be used downstream of 'find_coCREs'
}
